all tests working
