+++ .github/workflows/publish.yml
@@ -30,7 +30,7 @@ jobs:
       # Step 4: Run tests
       - name: Run tests
         run: |
-          pytest
+          pytest tests
 
   deploy:
     runs-on: ubuntu-latest
