+++ tests/pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
-testpaths = tests
 pythonpath = src
+filterwarnings =
+    ignore:Unverified HTTPS request
